Tailored analysis tables

Tailored analysis tables

Usage

tailored_patientsurvival(stcs)

tailored_organsurvival(stcs)

tailored_transplantationsurvival(stcs)

tailored_psq(stcs, silent = FALSE)

Arguments

stcs

A list containing the STCS data frame.

silent

lgl. Specify if warning should be returned.

Value

A data frame.

Details

tailored_organsurvival(): PK: organkey. Main dates to to compute follow-up time of organs.

1. patientkey: Identifier of patients. Source: Organ.

2. donorkey: Identifier of donors. Source: Organ.

3. soaskey: Identifier of transplantations. Source: Organ.

4. organkey: Identifier of organs. Source: Organ.

5. tpxdate: Date of transplantation. Source: Organ.

6. glo_date: Graft loss date. Source: Graftloss.

7. pnf_date: Date of PNF (tpxdate when dgf is "PNF"). Source: Organ.

8.-16. last_[...]_toggle_date: Assessment date of the last mentioned filled toggle (keep: "Yes", "No", removed: "Missing", NA). Source: OrganLongitudinal.

17. last_complete_organlongitudinal_crf_date: Assessment date of the latest crf_status of the organ assessment defined as "Complete". Source: OrganLongitudinal.

tailored_patientsurvival(): PK: patientkey. Main dates to to compute follow-up time of patients.

1. patientkey: Identifier of patients. Source: Patient.

2. enrollment_date: Enrollment date. Source: Patient.

3. deathdate: Death date. Source: Stop.

4. first_dropoutdate: First dropout date. Source: Stop.

5. first_dropoutdateaccuracy: Accuracy of the first dropout date. Source: Stop.

6. last_activedropoutdate: Last dropout date without backstcs. Source: Stop.

7. last_activedropoutdateaccuracy: Accuracy of the last dropout date without backstcs. Source: Stop.

8. last_patlongkeys: Identifier of the last patient-assessment. Source: PatientLongitudinal.

9. last_assdate: Date of the last patient-assessment. Source: PatientLongitudinal.

10.-17. last_[...]_toggle_date: Assessment date of the last mentioned filled toggle (keep: "Yes", "No", removed: "Missing", NA). Source: PatientLongitudinal.

18. extraction_date: Extraction date. Source: Admin.

19. last_complete_patientlongitudinal_crf_date: Assessment date of the latest crf_status of the patient assessment defined as "Complete". Source: PatientLongitudinal.

20. last_complete_stop_crf_date: Date (latest between: deathdate, dropoutdate, lastalivedate, backstcsdate) of the latest crf_status defined as "Complete". Source: Stop.

21. last_consent_withdrawal_date: Last consent date when the its consent status is "Withdrawal". Source Patient.

22. confidental data: Optional.

23. last_complete_psq_crf_date: Optional. Date of the latest crf_status of PSQ defined as "Complete". Source: PSQ.

tailored_psq(): PK: patlongkey. All PSQ forms in wide format.